+++ src/libtest/stats.rs
@@ -298,8 +298,8 @@ fn percentile_of_sorted(sorted_samples: &[f64], pct: f64) -> f64 {
     let lrank = rank.floor();
     let d = rank - lrank;
     let n = lrank;
-    let lo = sorted_samples[n];
-    let hi = sorted_samples[n+1];
+    let lo = sorted_samples[n as usize];
+    let hi = sorted_samples[n as usize + 1];
     lo + (hi - lo) * d
 }
