+++ src/test/run-pass/cast-region-to-uint.rs
@@ -10,5 +10,5 @@
 
 pub fn main() {
     let x: isize = 3;
-    println!("&x={:x}", (&x as *const isize));
+    println!("&x={:x}", (&x as *const isize as usize));
 }
